+++ lib/src/game/base_game.dart
@@ -1,9 +1,6 @@
 import 'dart:ui';
 
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/widgets.dart' hide WidgetBuilder;
+import 'package:meta/meta.dart';
 import 'package:ordered_set/comparing.dart';
 import 'package:ordered_set/ordered_set.dart';
 
@@ -11,6 +8,7 @@ import '../geometry/collision_detection.dart' as collision_detection;
 import '../components/component.dart';
 import '../components/mixins/collidable.dart';
 import '../components/mixins/has_collidables.dart';
+import '../components/mixins/draggable.dart';
 import '../components/mixins/has_game_ref.dart';
 import '../components/mixins/tapable.dart';
 import '../components/position_component.dart';
@@ -49,6 +47,12 @@ class BaseGame extends Game with FPSCounter {
       assert(
         this is HasTapableComponents,
         'Tapable Components can only be added to a BaseGame with HasTapableComponents',
+      );
+    }
+    if (c is Draggable) {
+      assert(
+        this is HasDraggableComponents,
+        'Draggable Components can only be added to a BaseGame with HasDraggableComponents',
       );
     }