+++ CHANGELOG.md
@@ -18,6 +18,8 @@
  - Added a collision detection system
  - Added geometrical shapes
  - Fix `SpriteAnimationComponent.shouldRemove` use `Component.shouldRemove`
+ - Add assertion to make sure Draggables are safe to add
+ - Add utility methods to the Anchor class to make it more "enum like"
 
 ## 1.0.0-rc6
  - Use `Offset` type directly in `JoystickAction.update` calculations